Btrfs: clear EXTENT_DEFRAG bits in finish_ordered_io
Before this, we use 'filled' mode here, ie. if all range has EXTENT_DEFRAG bits, go to clear it, but if the defrag range joins the adjacent delalloc range, then we'll leave EXTENT_DEFRAG bits until evicting inode.

This clear that bit if any was found within the ordered extent.

Signed-off-by: Liu Bo <bo.li.liu@oracle.com>